+++ apps/docs/stories/components/payouts-list.stories.tsx
@@ -33,6 +33,8 @@ export const Basic: Story = {};
 export const Styled: Story = {
   args: {
     style: {
+      'justifi-payouts-list::part(table-head)': {},
+      'justifi-payouts-list::part(table-head-row)': {},
       'justifi-payouts-list::part(table-head-cell)': {
         'background-color': '#fff',
         'font-weight': '600',
@@ -40,19 +42,43 @@ export const Styled: Story = {
         'text-transform': 'uppercase',
         'letter-spacing': '0.1em',
       },
+      'justifi-payouts-list::part(table-body)': {},
+      'justifi-payouts-list::part(table-row)': {},
+      'justifi-payouts-list::part(table-row):hover': {
+        'cursor': 'pointer',
+      },
+      'justifi-payouts-list::part(table-row-even)': {},
+      'justifi-payouts-list::part(table-row-odd)': {},
       'justifi-payouts-list::part(table-cell)': {
         'background-color': 'transparent',
         'font-weight': 'normal',
         'font-size': '0.8rem',
       },
+      'justifi-payouts-list::part(loading-state-cell)': {},
+      'justifi-payouts-list::part(loading-state-spinner)': {
+        'color': '#ccc',
+      },
+      'justifi-payouts-list::part(error-state)': {},
+      'justifi-payouts-list::part(empty-state)': {},
       'justifi-payouts-list::part(pagination-bar)': {
         'background-color': '#fff',
-        'border-bottom': 'none'
+        'border-bottom': 'none',
       },
-      'justifi-payouts-list::part(table-row):hover': {
-        'background-color': 'rgba(0, 0, 0, 0.05)',
-        'cursor': 'pointer'
-      }
+      'justifi-payouts-list::part(page-button)': {
+        'border': 'none',
+        'background-color': 'transparent',
+        'text-transform': 'uppercase',
+        'font-weight': 'normal',
+        'font-size': '0.8rem',
+      },
+      'justifi-payouts-list::part(page-button-disabled)': {
+        'opacity': '0.5',
+        'cursor': 'not-allowed',
+      },
+      'justifi-payouts-list::part(page-arrow)': {
+        'display': 'none',
+      },
+      'justifi-payouts-list::part(page-button-text)': {},
     }
   }
 };