+++ packages/webcomponents/src/components/pagination-menu/pagination-menu.tsx
@@ -18,33 +18,31 @@ export class PaginationMenu {
 
   render() {
     return (
-      <Host>
+      <Host exportparts={`page-arrow,page-button,page-button-disabled,page-button-text`}>
         <nav aria-label="Table pagination" class="d-flex justify-content-end gap-3">
           <ul class="pagination m-0">
-            <li
-              onClick={(e) => {
-                e.preventDefault();
-                this.paging.handleClickPrevious(this.paging.start_cursor);
-              }}
-              part={`previous-button${this.paging.has_previous ? '' : ' button-disabled'}`}
-              class={`page-item ${this.paging.has_previous ? '' : ' disabled'}`}
-            >
-              <a href="#" class="page-link">
-                <span class="me-1" part="arrow arrow-left">&laquo;</span>
-                <span part="previous-button-text">Previous</span>
+            <li class={`page-item ${this.paging.has_previous ? '' : ' disabled'}`}>
+              <a href="#"
+                class="page-link"
+                onClick={(e) => {
+                  e.preventDefault();
+                  this.paging.handleClickPrevious(this.paging.start_cursor);
+                }}
+                part={`page-button${this.paging.has_previous ? '' : ' page-button-disabled'}`}>
+                <span class="me-1" part="page-arrow">&laquo;</span>
+                <span part="page-button-text">Previous</span>
               </a>
             </li>
-            <li
-              onClick={(e) => {
-                e.preventDefault();
-                this.paging.handleClickNext(this.paging.end_cursor);
-              }}
-              part={`next-button${this.paging.has_next ? '' : ' button-disabled'}`}
-              class={`page-item ${this.paging.has_next ? '' : ' disabled'}`}
-            >
-              <a href="#" class="page-link">
-                <span part="next-button-text">Next</span>
-                <span class="ms-1" part="arrow arrow-right">&raquo;</span>
+            <li class={`page-item ${this.paging.has_next ? '' : ' disabled'}`}>
+              <a href="#"
+                class="page-link"
+                onClick={(e) => {
+                  e.preventDefault();
+                  this.paging.handleClickNext(this.paging.end_cursor);
+                }}
+                part={`page-button${this.paging.has_next ? '' : ' page-button-disabled'}`}>
+                <span part="page-button-text">Next</span>
+                <span class="ms-1" part="page-arrow">&raquo;</span>
               </a>
             </li>
           </ul>